fixes bugs with detecting changes in the board
also adds some filtering methods and lets us check to see if the board is valid using a different setup without actually having the pieces be there. This also adds a method that combines the addlist, the ignorelist, and the board list prioritizing the add pieces list then the ignore locs list, then the board list.