adds delete method
and it adds a pawning method and a castling method. This sets the notation for generating the moves and displaying them to the user. However moves are not made. You can still use setLoc and setMoveCount to get around some problems with moving, but you need to call removePieceAt at the right time. Pawning takes care of this, but basic moves do not for the moment.